// src/types/enums/typography-type.enum.ts
export enum TypographyType {
	HEADING_1 = 'HEADING_1',
	HEADING_2 = 'HEADING_2',
	HEADING_3 = 'HEADING_3',
	HEADING_4 = 'HEADING_4',
	HEADING_5 = 'HEADING_5',
	HEADING_6 = 'HEADING_6',
	BODY_1 = 'BODY_1',
	BODY_2 = 'BODY_2',
	PARAGRAPH = 'PARAGRAPH',
	BUTTON = 'BUTTON',
	CAPTION = 'CAPTION',
	OVERLINE = 'OVERLINE',
}